refactor: change EventOrdinal type from short to int across models and database scripts

+++ src/Outbox.Core/Models/DeadLetteredMessage.cs
@@ -12,7 +12,7 @@ public sealed record DeadLetteredMessage(
     byte[] Payload,
     string PayloadContentType,
     DateTimeOffset EventDateTimeUtc,
-    short EventOrdinal,
+    int EventOrdinal,
     int RetryCount,
     DateTimeOffset CreatedAtUtc,
     DateTimeOffset DeadLetteredAtUtc,
