+++ src/Outbox.Core/Models/OutboxMessageContext.cs
@@ -16,7 +16,7 @@ public sealed class OutboxMessageContext
     public string PartitionKey => _original.PartitionKey;
     public string EventType => _original.EventType;
     public DateTimeOffset EventDateTimeUtc => _original.EventDateTimeUtc;
-    public short EventOrdinal => _original.EventOrdinal;
+    public int EventOrdinal => _original.EventOrdinal;
     public int RetryCount => _original.RetryCount;
     public DateTimeOffset CreatedAtUtc => _original.CreatedAtUtc;
